Refactor, rename js to jsx and clean code
