--- src/app/atencion-al-cliente/page.jsx
@@ -1,9 +1,0 @@
-export default function Page() {
-    return (
-        <div className="min-h-screen flex justify-center items-center ">
-            <h1>
-                Atencion al cliente
-            </h1>
-        </div>
-    )
-}
