+++ src/app/como-denunciar/page.jsx
@@ -1,8 +1,8 @@
+import Image from "next/image";
 import ComoDenunciarFirstSection from "@/components/ComoDenunciar/ComoDenunciarFirstSection";
 import ComoDenunciarMainSection from "@/components/ComoDenunciar/ComoDenunciarMainSection";
 import ComodenunciarSecondSection from "@/components/ComoDenunciar/ComoDenunciarSecondSection";
 import ComoDenunciarThirdSection from "@/components/ComoDenunciar/ComoDenunciarThirdSection";
-import Image from "next/image";
 import bgimage1 from "../../../public/images/como-denunciar-image1.png";
 import bgimage2 from "../../../public/images/como-denunciar-image2.png";
 import image3 from "../../../public/images/como-denunciar-image3.png";
