+++ src/app/foro/[id]/page.jsx
