+++ src/app/registro/page.jsx
