+++ src/components/ComoDenunciar/ComoDenunciarSecondSection.jsx
@@ -1,5 +1,3 @@
-import Image from "next/image";
-
 export default function ComodenunciarSecondSection() {
   return (
     <>
